Updated rpcs3 to C++ flatimage

+++ rpcs3/build-arch.sh
@@ -8,6 +8,8 @@
 # @description : 
 ######################################################################
 
+#shellcheck disable=2016
+
 set -e
 
 function msg()
@@ -35,28 +37,38 @@ function fetch_rpcs3()
 
   # Remove image
   rm "$BUILD_DIR/$appimage_rpcs3"
+
+  # Move rpcs3 dir
+  mv "$BUILD_DIR"/squashfs-root/usr rpcs3
+
+  # Export rpcs3 dir location
+  export RPCS3_DIR="$BUILD_DIR"/rpcs3
+
+  # Remove squashfs-root
+  rm -rf ./squashfs-root
 }
 
 function fetch_flatimage()
 {
   msg "${BUILD_DIR:?BUILD_DIR is undefined}"
 
-  # Fetch container
-  if ! [ -f "$BUILD_DIR/arch.tar.xz" ]; then
-    wget "$(wget -qO - "https://api.github.com/repos/ruanformigoni/flatimage/releases/latest" \
-      | jq -r '.assets.[].browser_download_url | match(".*arch.tar.xz$").string')"
+  if [[ -n "$1" ]]; then
+    cp "$1" "$IMAGE"
+  else
+    # Fetch container
+    if ! [ -f "$BUILD_DIR/arch.tar.xz" ]; then
+      wget "$(wget -qO - "https://api.github.com/repos/ruanformigoni/flatimage/releases/latest" \
+        | jq -r '.assets.[].browser_download_url | match(".*arch.tar.xz$").string')"
+    fi
+
+    # Extract container
+    rm -f "$IMAGE"
+
+    tar xf arch.tar.xz
   fi
 
-  # Extract container
-  rm -f "$IMAGE"
-
-  tar xf arch.tar.xz
-
-  # FIM_COMPRESSION_LEVEL
-  export FIM_COMPRESSION_LEVEL=6
-
-  # Resize
-  "$IMAGE" fim-resize 3G
+  # Enable network
+  "$IMAGE" fim-perms set network
 
   # Update
   "$IMAGE" fim-root fakechroot pacman -Syu --noconfirm
@@ -76,44 +88,25 @@ function fetch_flatimage()
   "$IMAGE" fim-root fakechroot pacman -S libappindicator-gtk3 \
     lib32-libappindicator-gtk3 --noconfirm
 
-  # Compress self
-  "$IMAGE" fim-compress
+  # Commit changes
+  "$IMAGE" fim-commit
 }
 
 
 function compress_rpcs3()
 {
   msg "${BUILD_DIR:?BUILD_DIR is undefined}"
-
   # Copy rpcs3 runner
-  cp "$SCRIPT_DIR"/boot.sh "$BUILD_DIR"/squashfs-root/usr/boot
-
+  cp "$SCRIPT_DIR"/boot.sh "$RPCS3_DIR"/boot
+  # Create layer dirs
+  mkdir -p ./root/opt
+  mkdir -p ./root/home/rpcs3/.config
+  # Move rpcs3 to layer dir
+  mv "$RPCS3_DIR" ./root/opt
   # Compress rpcs3
-  "$IMAGE" fim-exec mkdwarfs -i "$BUILD_DIR"/squashfs-root/usr -o "$BUILD_DIR/rpcs3.dwarfs"
-}
-
-function hooks_add()
-{
-  msg "${IMAGE:?IMAGE is undefined}"
-  msg "${SCRIPT_DIR:?SCRIPT_DIR is undefined}"
-
-  "$IMAGE" fim-hook-add-pre "$SCRIPT_DIR"/hook-rpcs3.sh
-}
-
-function configure_flatimage()
-{
-  msg "${IMAGE:?IMAGE is undefined}"
-
-  # Set default command
-  # shellcheck disable=2016
-  "$IMAGE" fim-cmd '"$FIM_BINARY_RPCS3"'
-
-  # Set perms
-  "$IMAGE" fim-perms-set wayland,x11,pulseaudio,gpu,session_bus,input,usb
-
-  # Set up HOME
-  #shellcheck disable=2016
-  "$IMAGE" fim-config-set home '"${FIM_DIR_BINARY}"'
+  "$IMAGE" fim-layer create ./root "$BUILD_DIR/rpcs3.dwarfs"
+  # Remove uncompressed files
+  rm -rf ./root
 }
 
 function package()
@@ -154,11 +147,30 @@ function main()
   # Container file path
   export IMAGE="$BUILD_DIR/arch.flatimage"
 
+  # Fetch rpcs3
   fetch_rpcs3
-  fetch_flatimage
+
+  # FlatImage
+  if [[ "$1" = "--flatimage" ]] && [[ -n "$2" ]]; then
+    fetch_flatimage "$2"
+  else
+    fetch_flatimage
+  fi
+
+  # Create novel layer
   compress_rpcs3
-  hooks_add
-  configure_flatimage
+
+  # Set variables
+  "$IMAGE" fim-env set 'HOME=/home/rpcs3' \
+    'PATH="/opt/rpcs3/bin:$PATH"' \
+    'FIM_BINARY_RPCS3="/opt/rpcs3/boot"'
+
+  # Set default command
+  "$IMAGE" fim-boot '/opt/rpcs3/boot'
+
+  # Set perms
+  "$IMAGE" fim-perms set media,audio,wayland,xorg,dbus_user,dbus_system,udev,usb,input,gpu,network
+
   package
 }
